Cosmetics

+++ src/cosine.cc
@@ -24,7 +24,7 @@
 
 #include "utils.hh"
 
-n4::actions *create_actions(u16 nphot, const sim_config& s, const geometry_config& g) {
+n4::actions *create_actions(u32 nphot, const sim_config& s, const geometry_config& g) {
 
   std::unique_ptr<random_position> pos;
 
@@ -45,7 +45,6 @@ n4::actions *create_actions(u16 nphot, const sim_config& s, const geometry_confi
     -> dir(std::make_unique<n4::random::direction>())
     -> pol(std::make_unique<n4::random::direction>())
     -> ene(std::make_unique<lxe_scintillation>())
-    // -> fix_ene(7.21 * eV)
     ;
 
   return  (  new n4::        actions{ gen })
